+++ govcd/catalog_test.go
@@ -427,7 +427,7 @@ func (vcd *TestVCD) Test_CatalogDeleteMediaImage(check *C) {
 
 	AddToCleanupList(itemName, "mediaImage", vcd.org.Org.Name+"|"+vcd.vdc.Vdc.Name, "Test_UploadCatalogMediaImage")
 
-	mediaItem, err := vcd.vdc.FindMediaImage(itemName)
+	mediaItem, err := vcd.vdc.QueryMediaImage(itemName, vcd.config.VCD.Catalog.Name)
 	check.Assert(err, IsNil)
 	check.Assert(mediaItem, Not(Equals), MediaItem{})
 
@@ -436,7 +436,7 @@ func (vcd *TestVCD) Test_CatalogDeleteMediaImage(check *C) {
 	err = task.WaitTaskCompletion()
 	check.Assert(err, IsNil)
 
-	mediaItem, err = vcd.vdc.FindMediaImage(itemName)
+	mediaItem, err = vcd.vdc.QueryMediaImage(itemName, vcd.config.VCD.Catalog.Name)
 	check.Assert(err, IsNil)
 	check.Assert(mediaItem, Equals, MediaItem{})